Don't send Connection:close
Since I'm having enormous difficulty getting nginx to send Content-Length
or Transfer-Encoding:chunked when TPT sends this, and TPT requires one of
those headers for the HTTP request to work correctly.

Other programs don't seem to need one of those headers, so there's probably
a better way of fixing this.