Reorganised air sim
Threading and locking for air sim needs some more work (would be easier if simulation and UI were decoupled and in separate threads). At the moment, it's synchronous. In theory, air sim could run between the end of the main particle loop in one frame and the start of the main particle loop in the next frame.

Non-particle-order-dependent ambient heat will need some careful thought.